move config to common/config
