+++ product/.eslintrc
@@ -1,5 +1,5 @@
 {
   "extends": [
-    "../.eslintrc"
+    "../common/config/.eslintrc"
   ]
 }
